BlobProject/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# Azure and Cloud Storage
azure-storage-blob==12.17.0
django-storages==1.14.2

# Database (optional, choose based on your database)
psycopg2-binary==2.9.9  # For PostgreSQL
# Or keep SQLite for development

# Authentication and Security
django-cors-headers==4.2.0
whitenoise==6.6.0
python-dotenv==1.0.0

# File Handling
Pillow==10.1.0  # Image processing

# API and Serialization
markdown==3.5.1  # Optional: Markdown support for DRF
django-filter==23.3  # Optional: Filtering for DRF

# Development and Debugging
ipython==8.17.2
pygments==2.16.1

# Performance and Monitoring (optional)
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
model-bakery==1.14.0


# Optional but recommended
requests==2.31.0